refactor: move all validation and submit logic to the appropriate forms
